workingfiles/requirements.txt
matplotlib==3.5.2
numpy==1.22.1
Pillow==9.2.0
tensorflow==2.9.1
tensorflow_gpu==2.9.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
